+++ ReleaseNotes.md
@@ -1,4 +1,5 @@
 
+* 3.0.1 - OLED profile for AZDeliveryOLED128x64 added. Addresses moved to profile.
 * 3.0.0 - Added OLED support. Net40 version with code links to netstandard files.
 * 2.1.0 - Moved Display namespace to separate library. Added some OLED support.
 * 2.0.0 - Removed v1 DigitalIo usage.
@@ -6,5 +7,5 @@
 * 1.2.0 - Added PWM interfaces and soft PWM implementation (Core only)
 * 1.1.2 - Nuget package logo changed
 * 1.1.1 - Added Input and Output convienience classes, version info and help display
-* 1.1.0 - Splittet to full framework and core versions
+* 1.1.0 - Splitted to full framework and core versions
 * 1.0.0 - Initial release